Started refactoring shared test and example code into a module.
Cargo doesn't seem to have a way to say "this module uses dev-dependencies and is only visible to tests and examples", so it's awkwardly behind a magic feature.

The other option was to put the tests and examples in a separate crate. This would be nicer for library users as the main crate stays clean. But it's pain for development, and tests could no longer access library internals.